+++ NEWS.md
@@ -1,5 +1,7 @@
 # hardhat (development version)
 
+* Added a new vector class called `quantile_pred()` to house predictions made from a quantile regression model (tidymodels/parsnip#1191, @dajmcdon). 
+
 # hardhat 1.4.0
 
 * Added `extract_postprocessor()` generic (#247).